Django-Middleware-0x03/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
environ==1.0
Faker==37.3.0
mysql==0.0.3
mysql-connector-python==9.3.0       
mysqlclient==2.2.7
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
rest_framework_simplejwt==0.0.2